cleanup; refactor upper filesystem client; optimize directory reads dispatching;
