+++ futurecommander_gui/app_web/entry/component.js
@@ -29,6 +29,7 @@ module.exports = {
         };
 
         if(this.store.getState().is_open){
+            console.log('SPOIL');
             this.spoil();
         }
     },
